requirements.txt
langchain==0.0.275
fastapi==0.103.0
uvicorn==0.23.2
psycopg2==2.9.7
openai==0.27.10
tiktoken==0.4.0
python-dotenv==1.0.0

